This fixes a problem with Vanity when running inside a functional test.
The rails.rb file is checking for a response object and, if one exists, it generates a unique "vanity_identity" value and sets it to a cookie. The problem is that when the test is setup (before a response), an identity is created called "test" where the alternative index is set. Once the request is made, and since a response now exists, Vanity is does not look at the right key/value pair so it generates a new"vanity_identity" value and assigns a different, unexpected index value there.

This fix adds a check that detects a response object and, if one exists, that its not an instance of TestResponse. This ensures that the identity will remain "test" throughout the test case.