This fixes a bug with Vanity when running inside a functional test.
The choose() method choose a value for a particular experiment by picking an alternative out of a collection of alternatives. This is done by generating an Array index value based on what alternative was previously shown to the user or by using some logic to determine which alternative should be shown.

The problem occurs when collecting is disabled. The index value is never set so it will always choose alternative 0, regardless of which test has been explicitly selected in the test.